backend/requirements.txt
fastapi==0.122.1
uvicorn[standard]==0.38.0
tortoise-orm==0.25.1
asyncpg==0.31.0
pydantic==2.12.5
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
aerich==0.9.2
email-validator==2.2.0

